Fix JAR sig verification with multiple digests
This fixes a bug in ApkVerifier (and thus also in "apksigner verify")
where it would incorrectly reject APKs whose MANIFEST.MF contains
multiple digests for the same entry. Verification would fail with

ERROR: SHA-256 digest of AndroidManifest.xml does not match the digest
specified in META-INF/MANIFEST.MF. Expected:
<m5RN5FDhPTSIGRMQ0uVUXeEI+J0i51uoGsXVFFZGbnQ=>, actual:
<2jlj7l5rSw0yVb/vlWAYkK/YBwk=>

demonstrating that due to this bug ApkVerifier was comparing a SHA-1
digest to a SHA-256 digest.

This commit also adds unit tests for this issue and related issued.

Test: bazel test ...
Test: gradlew test
Bug: 38497270
Change-Id: Ifb7d44ffd59e6c2239d1d73b02e04e99a20dc09a