User-friendlier error when JAR sig alg unsupported
When reporting that an APK's JAR signature could not be verified
because it uses an unsupported digest or signature algorithm,
apksig/apksigner reported the OIDs of these two algorithms because
this is what's stored in the actual signature. This causes confusion
for users.

This commit thus adjusts this error message to use user-friendlier
algorithm names instead of OIDs. For example, SHA-256 instead of
2.16.840.1.101.3.4.2.1.

This commit also makes the list of API Levels look more readable.

Test: apksigner verify --min-sdk-version=14 \
          src/test/resources/com/android/apksig/v1-only-with-rsa-pkcs1-sha256-1.2.840.113549.1.1.1-2048.apk
      Produces user-friendlier error without OIDs:
      ERROR: JAR signer CERT.RSA: JAR signature META-INF/CERT.RSA
      uses digest algorithm SHA-256 and signature algorithm RSA which
      is not supported on API Level(s) 14-17 for which this APK is being
      verified
Test: bazel test ...
Test: gradlew test
Bug: 63525618
Change-Id: If240b4199e0757f89d3eed0b71b20fbb8104b2c7