Add --pass-encoding parameter to apksigner
When keytool creates a KeyStore or key which is protected with a
password containing non-ASCII characters, keytool may encode the
password using the console's encoding or the JVM default encoding
instead of using the password verbatim, in its Unicode form. In
apksigner 0.5 support was added to handle this issue by having
apksigner try several variants of the password: verbatim/Unicode,
encoded using console encoding, and encoded using JVM default
encoding. Unfortunatly, there is no public API to obtain the
console's encoding. apksigner 0.5 cheated by accessing
Console.encoding field via Reflection API. This is no longer
advisable in Java 9 because access to such private API is brittle
and now leads to JVM warnings spewing on the standard output.

To make apksigner compatible with Java 9, apksigner's behavior is
being changed by this commit. apksigner no longer attempts to detect
console encoding on Java 9. Those users which use non-ASCII password
keystores/keys created by keytool in environments where console
encoding is not the same as the JVM default encoding, such as English
Windows where console encoding is ibm437 and JVM default encoding is
windows-1252, will need to pass in the new --pass-encoding <encoding>
parameter (e.g., --pass-encoding ibm437) to tell apksigner to try
a password variant encoded using that encoding, in addition to trying
the Unicode and JVM default encoding variants of the password.

The --pass-encoding parameter enables apksigner to work with non-ASCII
password keystores/keys created on different OSes and/or in other
locales than the one in which apksigner is running. For example, this
enables apksigner on OSX and Linux (where the console and JVM default
encoding is UTF-8) to use a non-ASCII password keystore created by
keytool on English Windows, where the resulting password is encoded
using the ibm437 or windows-1252 encoding.

Test: Use keytool to create two JKS keystores protected using password
      "ab¡äю1" (one where the password is read from the console, the
      other where the password is provided on the command-line). Use
      apksigner to sign an APK using these keystores, using Java 8 and
      Java 9 (revision 181), providing the password via all four
      methods supported by apksigner. Do this on Windows 10 (IBM437
      console encoding, windows 1252 file/JVM default encoding, OSX
      (UTF-8), and Linux (UTF-8).
Bug: 37135737
Bug: 37137869
Change-Id: Ia6fb0706e4bb3f0fd968373d113d457e8c6260c8