Public API for locating APK Signing Block
This exposes as public API the standard way to locate the APK Signing
Block in an APK: ApkUtils.findApkSigningBlock.

The functionality was moved from V2SchemeVerifier.findApkSigningBlock.
It is fine to move this functionality from V2SchemeVerifier because
APK Signing Block is not specific to APK Signature Scheme v2 -- the
block is a generic mechanism for adding additional signatures to the
APK.

Test: bazel test ...
Change-Id: If01619556ce543fec63786f5105921d0e1c0fc0d