Refactor non-v2-specific components into common area for v3 use.
Following the addition of platform support for APK Signature Scheme v3
(commit: 67096e08a72beea85979a3aa9fc5b376b2c2b5ad), apksig will need to
be modified to support this new version.  Refactor apksig to split apart
the components which are necessary for using the APK Signing Block from
those that are specific to the v2 scheme inside of it.

Bug: 64686581
Test: bazel test ...
Change-Id: I6ec3f888ebd155c1f2e262ee7d849d752f1afe88