Ensure v1/v2 signer matches root signing history.
In the event of APK Signature Scheme v3 signing certificate
changes, it is possible to maintain compatibility with previous
platforms by signing with a v1 and/or v2 signature as well.  In
this event, though, the v1/v2 signer should be be the oldest one in
the signing certificate history.  Enforce this to prevent v1/v2
signing with the wrong signing certificate when rotation has occurred.

Bug: 74074170
Test: bazel test ...
Change-Id: If06700490a3f4332e710ff13a9fc7d38789af6bd