Wrap JKS Certs in GuaranteedEncodedFormX509Certificates
In order to sign an APK with a rotated signing certificate the signer
lineage must be specified. The certificates in the signer lineage are
wrapped in GuaranteedEncodedFormX509Certificates, but the certificates
from the signing keystores are kept as X509Certificates. This causes
any invocation of the equals method against the lineage object to
fail when comparing against the signing certificates as
GuaranteedEncodedFormX509Certificate#equals will return false if the
other object is not a GuaranteedEncodedFormX509Certificate.

Fixes: 109670495
Test: apksigner sign --ks android.jks --next-signer --ks android2.jks
  --in app-release-unsigned-aligned.apk --lineage android-lineage.out
  --out app-release-signed.apk

Change-Id: Icf794f108c8d8b348fec17c2f4c9f668deb2c8ab