Reencode negative modulus in APK signing block to allow app updates
The V2 and V3 signature schemes verify the public key from the signature block
against the public key in the certificate that acts as the signature / identity for the
app. If the certificate is incorrectly encoded with a negative modulus (the modulus
has a leading 1 in the encoding without a preceding zero byte to encode the integer
as positive), then the V2 / V3 signature verification will fail as the public key will be
reencoded with a positive modulus which will not match the key in the signature block.
This change adds support for this improper encoding by reencoding the modulus for
the public key that is written in the signature block to ensure the public key verification
is successful on device.

Bug: 80184678
Test: Manually verified a certificate with a negative modulus could be used to sign and
      install APKs with the V2 and V3 signature schemes.

Change-Id: I24df8e30ed52a8034e75bd8f72cbd76209bb0db7