Change apksig API to allow disabling V4 error reporting
During cleanup of testV4State_disableV3EnableV4_fails in
ApkSignerTest I noticed the setV4SigningRequested may be confusing
to clients of apksig since the majority will consider
setV4SigningEnabled as requesting V4 signing. This commit modifies
the apksig API to treat setV4SigningEnabled as the caller requesting
V4 signing and modifies setV4SigningRequested to
setV4ErrorReportingEnabled to indicate the behavior of this method
on the signing process.

Bug: 153123157
Test: gradlew test
Change-Id: I4b1679eea413312249a078a6c8edd39a01d7f273