Use min SDK or default 1 when target SDK not specified
The apksig target SDK version minimum signature scheme check uses the
targetSdkVersion of the uses-sdk element to determine the minimum
signature scheme version to verify. If this attribute or the uses-sdk
element is not specified in the APK then the verifier should behave as
the platform and use the value of the minSdkVersion as the target or
a default value of 1 is neither is specified.

Bug: 148313868
Test: gradlew test
Change-Id: I79f6f80870c402fc603a7e01f9d76afaf0d1c3c3