Add support for Provider#configure in JDK 9+
Starting in JDK 9 the Provider single-arg configuration constructor
has been replaced with a #configure(String) method to be invoked
after the no-arg constructor. This commit will first attempt the
original behavior to support systems using an older JDK; if the
single-arg configuration constructor is not available then the new
API will be used.

Bug: 173837265
Test: gradlew test
Change-Id: I22d1143e76d1552d59e3cd6fadef66d9001e8009