Let pinlist.meta file signed for V1 signing.
The pinlist.meta generation, which was implemented at b/65316207,
didn't support V1 signing. This was not a problem because,
it seems that, the V1 signature isn't used for building
Google Camera APK. (The only known user of the pinlist.meta generation.)

But that changed after Google Camera transitioned from APK to AAB
for the deployment: The AAB-generated APKs use V1 signing.
This problem was masked by other bug (b/174697187) and fixing it
unmasked the problem, blocked the app from being pushed to the
production (b/179297646).

This change makes the V1 signing works for the pinlist.meta file.

Test: Unit test.
Bug: 174697187
Bug: 179297646

Change-Id: I164579043bec804c2cbc2b9166fdacc6a04963f1