Add support for APK Signature Scheme v3.1 for rotation targeting
This commit adds support for a new v3.1 APK signature scheme that
allows APK signing key rotation to target T+. This version will still
default to using the v3.0 signing block for rotation, but once the
build system is updated to support the new rotation-min-sdk-version
option v3.1 will be enabled by default for all key rotations.

This commit also updates all of the golden APKs that use key rotation
because the minSdkVersion of the v3 signer is updated from 24 (the
first API level that supports the signature algorithm) to 28 (the
first API level that supports v3).

Bug: 192301300
Test: gradlew test
Change-Id: I49cc98ea803d18d53131a78be668921d58ac5f4b