Add support to v3.1 sig scheme to target dev release
When a new release is under development, its SDK version is set to
the SDK version of the previous release; during the majority of the
T development, the SDK version of a device running T will be S (31).
Since S- devices do not know about the v3.1 block ID, it is safe to
set the minSdkVersion of a v3.1 signer using a rotated key to 31;
T devices will recognize the new signature scheme and use the rotated
signer, but if the same APK were installed on an S device the v3.1
block would be ignored, and the original signer would be used. However
once T is released and U is using the SDK version of T, if rotation
needs to target U, just specifying the SDK version of T will not be
sufficient since this could then install on T and U devices. The
new rotation-targets-dev-release flag will allow a v3.1 signer to
target the active release under development; if this new flag is set
then the minSdkVersion of the v3.1 signer must be within range of the
device's SDK version and the system property ro.build.version.codename
must not be set to "REL".

Bug: 192301300
Test: gradlew test
Change-Id: Id6ca0b6d5db575f301cf715ebaceadfd1df19de0