Ensure only a single signer is written to V3 block
Android T introduced SDK version targeting for rotation on T+ by
allowing a caller to specify a rotation-min-sdk-version. If this
value is less than T, then the rotated signing key should be used in
the V3.0 signing block. However, a value X > P and < T can cause
multiple signers to be written to the V3 block, one targeting P -
X-1 with the original signing key, and the other targeting X and
later with the rotated key. Since SDK version targeting has not
been thoroughly tested on previous platform releases, this commit
will set the rotation-min-sdk-version to P when a value X < T
is provided to ensure the V3 block only contains the rotated
signer. This remains consistent with the documentation and will
ensure apps that have already rotated prior to T can specify
the version on which they rotated to see the original V3 behavior.

Bug: 202424396
Test: gradlew test
Change-Id: I00dae110b8f9c552e6cd5100491b0caa5e7138f8