Add V4 signing tests to verify V3.1 support
Android T introduced the V3.1 signature scheme to allow APK version
targeting for key rotation; this scheme allows a single APK to support
rotation on SDK version X for X >= Android T, while all previous
releases will only see and use the original signing key. This requires
the V4 signature to be updated to include both the rotated signing key
when installed on a device running X+ as well as the original signing
key for devices running a version < X. This commit adds two new tests
to ensure the V4 signature contains the expected signer both when
the V3.1 signature scheme is used as well as when only the V3.0
scheme is used for rotation. The V3.1 test is currently marked with
@Ignore and should be re-enabled once the V4 scheme supports V3.1.

Bug: 202011194
Test: gradlew test
Change-Id: I4a4c43553c9e1e72c7b6d97c6b716ad57728666a