Do not write rotation-min-sdk to v3.0 block in the absence of rotation
apksig allows a value to be set for rotation-min-sdk-version without
requiring a rotated signing key / lineage to be specified. This allows
a generic signing config to be used that supports both dev and prod
builds where the dev build is using the rotated signign key but the prod
build is still using only the original key with the
rotation-min-sdk-version set for both. This commit ensures that the
rotation stripping attribute is only written to the v3.0 signer if
a rotated signing key is used in the v3.1 signing block.

Bug: 202830422
Test: gradlew test
Change-Id: I3010d591f7be6113708b01034e46d11b0352095c