Document issue with non-default capabilities for shared signers
Prior to Android S, when multiple apps share a common signer in their
signing lineage with non-default capabilities, the platform would
assign the capabilities declared for this signer in one of the app's
lineage to the same signer in the lineage of the rest of the apps.
This commit updates the developer signing docs to ensure developers
are aware of this issue.

Bug: 201588245
Test: gradlew test
Change-Id: If3e6395efdfe1a6c926933d3d6ee3857df1d6841