Update minSdkVersion for v3.1 block to Sv2 API level (32)
During the development of a new platform, the SDK version of the most
recently finalized platform release is used. Initially T used the
SDK version of S (31), but recently Sv2 was finalized and the SDK
version was bumped to 32. In order for the v3.1 signing block to
be recognized on a device running T, the device SDK version must fall
within the bounds of the v3.1 signer's min / max SDK version. The
v3.1 signature scheme will still work on devices running T with the
new SDK version of 32 since apksig was using 31 as the min SDK
version, but this value is also used as the max SDK version for the
v3.0 signer. Since Sv2 is also using 32 as its API level, the max
SDK version of 31 written to the v3.0 signer block prevents the device
from recognizing a proper v3.0 signer. This commit updates the
API level used for the v3.1 signer block to 32 which will also update
the v3.0 signing block to use 32 as the max SDK version allowing
APKs signed with the v3.1 signature scheme targeting T for rotation
to properly install with the original signer on a device running Sv2
with the finalized SDK.

Fixes: 205551461
Test: gradlew test
Change-Id: I44d65c36adfea4d792ae97afa1aac6ddfd09bdd3