Resolve verification errors for v3.1 sigs and min-sdk-version 33
apksig is designed to behave as the platform would based on the min /
max SDK versions; when a min-sdk-version is not explicitly provided,
apksig will use the minSdkVersion from the APK's manifest. A device
running Android T or later will only verify the v3.1 signature if it
exists; apksig will behave the same for an APK signed with a v3.1
signature and a min-sdk-version of 33 or later. This commit resolves
the verification errors that were reported when apksig only verified
the v3.1 signature with a min-sdk-version of 33 by properly
recognizing the v3.1 signature as an extension of v3 and also
sufficient to meet the min v2 signature requirement for apps targeting
SDK version 30.

Bug: 208504694
Test: gradlew test
Change-Id: I28eadf4103358eb23cbc10880ac90c394e54e333