Ensure v3.0 signer returned when app min SDK equals rotation min SDK
Android T introduced the v3.1 signature scheme that allows apps to
target APK signing key rotation to an SDK version T+ and the original
signing key will be used for previous releases. Since releases under
development use the SDK version of the previously released SDK, the
v3.1 scheme also supports targeting a development release; in the
case of T, the v3.1 block will target Sv2's SDK version (32) during
development and the v3.0 block will use 32 as the max SDK version
for the block. However, if an app's minSdkVersion is set to 32, the
previous behavior would only return the v3.1 signer causing v3.0
signing to fail due to a missing signing config. This commit
updates the logic to continue parsing the signers if the v3.1's
minSdkVersion is equal to the app's minSdkVersion when supporting
a development release.

Bug: 230340743
Test: gradlew test
Change-Id: Ibd13f899c7686cd5db6e509b6a2ecca73992b732