Add option to apksigner to display PEM encoding of signing certs
The apksigner comand line tool currently has an option to print the
details about an app's signing certificate(s), but there was no way
for a user to obtain the actual signing certificate. This commit
adds a new option to display the PEM encoding of each signing
certificate to stdout that can be used when verifying an APK's
signature or lineage.

Fixes: 230795090
Test: Manually verified PEM output was displayed when requested.
Change-Id: Ib7051d6d0bbdd99b95d37c0a9ccf7e06ef737b29