Update v3.1 signing block to target T's finalized SDK 33
The v3.1 signature scheme was introduced in Android T, but during the
development of T, the SDK version was set to that of the previously
released platform, Sv2, at 32. In order to test the new signature
scheme on T, the v3.1 signing block had to use a min SDK version of
32 to ensure the block's SDK range would apply to T. Now that the
T SDK has been finalized with a version of 33, this new value
can be used for packages targeting T for rotation.

Note, this will not affect packages signed with the previous min
SDK version of 32 since that would still apply to this finalized
SDK version.

Bug: 232118308
Test: gradlew test
Change-Id: I0764227c70806494f9c0361eecd068e228b4171e